[267 тыс. ключей в секунду]: временный отказ от SIMD в SHA256; замена временных регистров на clobbered registers; чуть более быстрое и намного более понятное умножение

+++ test.h
@@ -10,7 +10,7 @@ struct Timer
 	
 	Timer();
 	unsigned long long stop(int iterations);
-	unsigned long long stop();
+	double stop();
 };
 
 bool check(unsigned char* array1, unsigned char* array2, int length);
